+++ CHANGELOG.md
@@ -58,6 +58,7 @@
 - PR #1952 consolidate libcudf public API headers in include/cudf
 - PR #1949 Improved selection with boolmask using libcudf `apply_boolean_mask`
 - PR #1956 Add support for nulls in `query()`
+- PR #XXXX Refactor "legacy" groupby to use libcudf `groupby_without_aggregations` function
 
 
 ## Bug Fixes